examples/retrieval_function_call/requirements.txt
erniebot
paddle-pipelines
paddlepaddle
gradio